Rename evaluator -> evaluators for consistency with new modules
