+++ packages/phoenix-evals/tests/phoenix/evals/functions/test_executor.py
@@ -7,7 +7,7 @@ from unittest.mock import AsyncMock, Mock
 
 import nest_asyncio
 import pytest
-from phoenix.evals.executor import (
+from phoenix.evals.executors import (
     AsyncExecutor,
     SyncExecutor,
     get_executor_on_sync_context,
